removed useSearchParams to resolve noindex
