+++ middleware/Dockerfile
@@ -24,4 +24,4 @@ USER appuser
 
 COPY app .
 EXPOSE 8000
-CMD uvicorn 'app.main:app' --host=0.0.0.0 --port=8000
+CMD uvicorn 'main:app' --host=0.0.0.0 --port=8000
